feat: add close menu functional
