--- src/assets/style/_fonts.scss
@@ -1,6 +1,0 @@
-@import 'https://fonts.googleapis.com/css2?family=Audiowide&family=Oxygen:wght@300;400;700&family=Teko:wght@300..700&display=swap';
-
-@font-face {
-  font-family: Sneakers;
-  src: url('../fonts/Snickers.otf') format('opentype');
-}
