--- src/assets/style/_mixins.scss
@@ -1,11 +1,0 @@
-@use './variables' as vars;
-
-@mixin link-dark {
-  color: vars.$dark-main-color;
-  text-decoration: none;
-
-  &:hover {
-    color: vars.$dark-hover-color;
-    transition: 0.5s ease-in-out;
-  }
-}
