--- src/assets/style/_variables.scss
@@ -1,5 +1,0 @@
-/* vars */
-
-$dark-main-color: #000;
-$dark-hover-color: #7b7b7b;
-$light-main-color: #fff;
