--- src/components/containers/about/AboutUs.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function AboutPage(): React.JSX.Element {
-  return <div>About Page</div>;
-}
