--- src/components/containers/about/about-us.scss
@@ -1,1 +1,0 @@
-/* about */
