--- src/components/containers/bucket/bucket.scss
@@ -1,1 +1,0 @@
-/* bucket */
