--- src/components/containers/main/Main.tsx
@@ -1,3 +1,0 @@
-export default function MainPage(): React.JSX.Element {
-  return <div>Main Page</div>;
-}
