--- src/components/containers/main/main.scss
@@ -1,1 +1,0 @@
-/* main */
