--- src/components/containers/products/product-info/product-info.scss
@@ -1,1 +1,0 @@
-/* product-info */
