--- src/components/containers/products/product-list/ProductList.tsx
@@ -1,3 +1,0 @@
-export default function ProductList(): React.JSX.Element {
-  return <div>ProductList</div>;
-}
