--- src/components/containers/products/product-list/product-list.scss
@@ -1,1 +1,0 @@
-/* product-list */
