--- src/components/containers/user/login/LoginPage.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function LoginPage(): React.JSX.Element {
-  return <div>LoginPage</div>;
-}
