--- src/components/containers/user/login/login-page.scss
@@ -1,1 +1,0 @@
-/* login */
