--- src/components/containers/user/profile/profile.scss
@@ -1,1 +1,0 @@
-/* profile */
