--- src/components/containers/user/registration/registration-page.scss
@@ -1,1 +1,0 @@
-/* registration */
