--- src/components/header/header.scss
@@ -1,1 +1,0 @@
-/* header */
