+++ src/components/header/login-menu/LoginMenu.tsx
@@ -1,4 +1,4 @@
-import { useLoginMenu } from '../../../core/state/stateLoginMenu';
+import { useLoginMenu } from '../../../core/stores/stateLoginMenu';
 import styles from './login-menu.module.scss';
 import { useEffect, useRef } from 'react';
 
@@ -7,35 +7,31 @@ export default function LoginMenu(): React.JSX.Element {
 
   const menu: React.RefObject<null | HTMLDivElement> = useRef(null);
   useEffect(() => {
-    changeClass();
-    if (isOpen) {
-      const closeMenu = (event: MouseEvent): void => {
-        if (menu.current instanceof HTMLDivElement && event.target instanceof Node) {
-          if (!menu.current.contains(event.target)) {
-            toggleStatus();
-          }
-        }
-      };
-
+    if (!isOpen) return;
+    const closeMenu = (event: MouseEvent): void => {
+      if (menu.current && event.target instanceof Node && !menu.current.contains(event.target)) {
+        toggleStatus();
+      }
+    };
+    const timeout = setTimeout(() => {
       document.addEventListener('click', closeMenu);
-      return () => document.removeEventListener('click', closeMenu);
-    }
-  });
-
-  const changeClass = (): string => {
-    return isOpen ? 'open' : 'close';
-  };
+    }, 0);
+    return () => {
+      clearTimeout(timeout);
+      document.removeEventListener('click', closeMenu);
+    };
+  }, [isOpen, toggleStatus]);
 
   return (
-    <div className={`${styles['login-menu']} ${styles[changeClass()]}`} ref={menu}>
+    <div className={`${styles['login-menu']} ${isOpen ? styles.open : styles.close}`} ref={menu}>
       <div className={styles['login-menu__head']}></div>
       <div className={styles['login-menu__title']}>
         <h3>Login please</h3>
       </div>
-      <button type="button" className="login-menu__login">
+      <button type="button" className="login-menu__login" onClick={() => toggleStatus()}>
         Login
       </button>
-      <button type="button" className="login-menu__register">
+      <button type="button" className="login-menu__register" onClick={() => toggleStatus()}>
         Register
       </button>
     </div>
