--- src/components/pages/loading/loading.scss
@@ -1,1 +1,0 @@
-/* loading */
