--- src/components/pages/not-found/404.scss
@@ -1,1 +1,0 @@
-/* 404 */
