--- src/components/ui/button/button.scss
@@ -1,1 +1,0 @@
-/* button */
