--- src/components/ui/input/input.scss
@@ -1,1 +1,0 @@
-/* input */
