+++ src/pages/about/AboutUs.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function AboutPage(): React.JSX.Element {
+  return <div>About Page</div>;
+}
