+++ src/pages/main/Main.tsx
@@ -1,0 +1,3 @@
+export default function MainPage(): React.JSX.Element {
+  return <div>Main Page</div>;
+}
