+++ src/pages/products/product-list/ProductList.tsx
@@ -1,0 +1,3 @@
+export default function ProductList(): React.JSX.Element {
+  return <div>ProductList</div>;
+}
