+++ src/pages/user/login/LoginPage.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function LoginPage(): React.JSX.Element {
+  return <div>LoginPage</div>;
+}
